DDLS-591 Move API to PHP 8.2
