+++ api/app/phpstan-baseline.neon
@@ -1036,11 +1036,6 @@ parameters:
 			path: src/Controller/Report/ClientBenefitsCheckController.php
 
 		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 2
-			path: src/Controller/Report/ClientBenefitsCheckController.php
-
-		-
 			message: "#^Call to an undefined method Doctrine\\\\ORM\\\\EntityRepository\\<App\\\\Entity\\\\Report\\\\Contact\\>\\:\\:findByReport\\(\\)\\.$#"
 			count: 1
 			path: src/Controller/Report/ContactController.php
@@ -1308,11 +1303,6 @@ parameters:
 		-
 			message: "#^Parameter \\#1 \\$createdBy of method App\\\\Entity\\\\Report\\\\Document\\:\\:setCreatedBy\\(\\) expects App\\\\Entity\\\\User, Symfony\\\\Component\\\\Security\\\\Core\\\\User\\\\UserInterface\\|null given\\.$#"
 			count: 2
-			path: src/Controller/Report/DocumentController.php
-
-		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 1
 			path: src/Controller/Report/DocumentController.php
 
 		-
@@ -2962,11 +2952,6 @@ parameters:
 
 		-
 			message: "#^Parameter \\#1 \\$formData of method App\\\\Factory\\\\UserResearchResponseFactory\\:\\:generateFromFormData\\(\\) expects array, mixed given\\.$#"
-			count: 1
-			path: src/Controller/UserResearchController.php
-
-		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
 			count: 1
 			path: src/Controller/UserResearchController.php
 
@@ -8496,11 +8481,6 @@ parameters:
 			path: src/Security/LoginRequestAuthenticator.php
 
 		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 2
-			path: src/Security/LoginRequestAuthenticator.php
-
-		-
 			message: "#^Parameter \\#1 \\$password of class Symfony\\\\Component\\\\Security\\\\Http\\\\Authenticator\\\\Passport\\\\Credentials\\\\PasswordCredentials constructor expects string, mixed given\\.$#"
 			count: 1
 			path: src/Security/LoginRequestAuthenticator.php
@@ -8543,11 +8523,6 @@ parameters:
 		-
 			message: "#^Method App\\\\Security\\\\RegistrationTokenAuthenticator\\:\\:setBruteForceKey\\(\\) has no return type specified\\.$#"
 			count: 1
-			path: src/Security/RegistrationTokenAuthenticator.php
-
-		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 3
 			path: src/Security/RegistrationTokenAuthenticator.php
 
 		-
@@ -10321,11 +10296,6 @@ parameters:
 			path: src/v2/Controller/OrganisationController.php
 
 		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 2
-			path: src/v2/Controller/OrganisationController.php
-
-		-
 			message: "#^Parameter \\#1 \\$string of function strtolower expects string, string\\|null given\\.$#"
 			count: 1
 			path: src/v2/Controller/OrganisationController.php
@@ -10663,11 +10633,6 @@ parameters:
 		-
 			message: "#^Parameter \\#1 \\$fromRequest of method App\\\\v2\\\\Fixture\\\\Controller\\\\FixtureController\\:\\:createNdr\\(\\) expects array, mixed given\\.$#"
 			count: 1
-			path: src/v2/Fixture/Controller/FixtureController.php
-
-		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
-			count: 7
 			path: src/v2/Fixture/Controller/FixtureController.php
 
 		-
@@ -11372,11 +11337,6 @@ parameters:
 
 		-
 			message: "#^Cannot access offset 'secondClientId' on mixed\\.$#"
-			count: 1
-			path: src/v2/Tools/Controller/ToolsController.php
-
-		-
-			message: "#^Parameter \\#1 \\$json of function json_decode expects string, resource\\|string given\\.$#"
 			count: 1
 			path: src/v2/Tools/Controller/ToolsController.php
