+++ api/app/tests/Behat/bootstrap/Common/FormTrait.php
@@ -141,79 +141,4 @@ trait FormTrait
 
         $this->assertFieldContains($field, $year);
     }
-
-    /**
-     * Fills in form field with specified id|name|label|value.
-     *
-     * @When /^(?:|I )fill in "(?P<field>(?:[^"]|\\")*)" with "(?P<value>(?:[^"]|\\")*)"$/
-     * @When /^(?:|I )fill in "(?P<field>(?:[^"]|\\")*)" with:$/
-     * @When /^(?:|I )fill in "(?P<value>(?:[^"]|\\")*)" for "(?P<field>(?:[^"]|\\")*)"$/
-     */
-    public function fillField($field, $value)
-    {
-        $driver = $this->getSession()->getDriver();
-        $field = $this->fixStepArgument($field);
-        $value = $this->fixStepArgument($value);
-
-        if ('.' != substr($field, 0, 1) && '#' != substr($field, 0, 1)) {
-            $field = '#'.$field;
-        }
-
-        if ('Behat\Mink\Driver\Selenium2Driver' == get_class($driver)) {
-            $this->scrollTo($field);
-
-            $javascript = <<<EOT
-            var field = $('$field');
-            var value = '$value';
-
-            $(':focus').trigger('blur').trigger('change');
-            var tag = field.prop('tagName');
-
-            if (field.prop('type') === 'checkbox' ||
-                field.prop('type') === 'radio')
-            {
-
-                field.prop('checked', true);
-
-            } else if (tag === 'SELECT') {
-
-                field.focus().val(value).trigger('change');
-
-            } else {
-                var pos = 0,
-                    length = value.length,
-                    character, charCode;
-
-                for (;pos < length; pos += 1) {
-
-                    character = value[pos];
-                    charCode = character.charCodeAt(0);
-
-                    var keyPressEvent = $.Event('keypress', {which: charCode}),
-                        keyDownEvent = $.Event('keydown', {which: charCode}),
-                        keyUpEvent = $.Event('keyup', {which: charCode});
-
-                    field
-                        .focus()
-                        .trigger(keyDownEvent)
-                        .trigger(keyPressEvent)
-                        .val(value.substr(0,pos+1))
-                        .trigger(keyUpEvent);
-
-                }
-            }
-
-EOT;
-
-            $this->getSession()->executeScript($javascript);
-        } else {
-            $elementsFound = $this->getSession()->getPage()->findAll('css', $field);
-
-            if (empty($elementsFound)) {
-                throw new \RuntimeException("Element $field not found");
-            }
-
-            $elementsFound[0]->setValue($value);
-        }
-    }
 }
