+++ api/app/tests/Integration/Controller/PreRegistrationControllerTest.php
@@ -8,13 +8,11 @@ use App\Tests\Integration\Fixtures;
 
 class PreRegistrationControllerTest extends AbstractTestController
 {
-    private static User $deputy1;
-    private static User $admin1;
-    private static User $deputy2;
-    private static $tokenAdmin;
-    private static $tokenDeputy;
-    private static $tokenProf;
-    private static $tokenPa;
+    private static ?string $tokenAdmin = null;
+    private static string $tokenDeputy;
+    private static string $tokenProf;
+    private static string $tokenPa;
+    private ?PreRegistration $c1;
 
     /**
      * clear fixtures.
@@ -29,10 +27,6 @@ class PreRegistrationControllerTest extends AbstractTestController
     public function setUp(): void
     {
         parent::setUp();
-
-        self::$deputy1 = self::fixtures()->getRepo('User')->findOneByEmail('deputy@example.org');
-        self::$admin1 = self::fixtures()->getRepo('User')->findOneByEmail('admin@example.org');
-        self::$deputy2 = self::fixtures()->createUser();
 
         if (null === self::$tokenAdmin) {
             self::$tokenAdmin = $this->loginAsAdmin();
